requirements.txt
tqdm
ptflops 
accelerate 
coloredlogs 
einops
scikit-learn
wandb
torchvision
pandas
pyarrow